+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## [unreleased]
 
+ * OpenID Connect Security Provider: Fixed the problem that authentication attempts could fail if the IDP redirected back
+   very quickly, but the session was not persisted in the store yet
  * Portal: Added the attribute *data-mr-app-name* to the default App wrapper to simplify end-2-end testing
  * Portal: The SSR route of Remote Apps will receive now also the path and the query parameters of the original request (fixes #102).
    The body of the POST request looks like this now: